Weather: Add Hourly Forecasts.
Now, we show next 4 hours forecasts
along with the current weather/temperature.

We fetch the hourly data we need to use a new API
provided by openweathermap.org which requires
latitude and longitude information which is
provided by a new API from opencagedata.
Environment variables for the same is update
as well.

Fixes: #1.